### 2025-11-17 #731 @paytonyau
OptSurvCutR: Validated Cut-point Selection for Survival Analysis
- Submitting Author Name: Payton Yau
- Submitting Author Github Handle: @paytonyau
- Repository: https://github.com/paytonyau/OptSurvCutR
- Version submitted: 0.1.8
- Submission type: Stats
- Badge grade: bronze
- Editor:  TBD
- Reviewers:  TBD

- Archive: TBD
- Version accepted: TBD
- Language: en
---

```
Package: OptSurvCutR
Title: Optimal Survival Cut-Point Discovery for Time-to-Event Analysis
Version: 0.1.9
    person(
        given = "Payton",
        family = "Yau",
        email = "tungon@gmail.com",
        comment = c(ORCID = "0000-0002-3283-0370")
    )
    Provides a robust 3-step workflow for optimal cut-point analysis in
    time-to-event (survival) data. Functions determine the optimal number
    of cut-points ('find_cutpoint_number()'), find their precise locations
    ('find_cutpoint()') using systematic or genetic algorithms (via 'rgenoud'),
    and validate stability via bootstrapping ('validate_cutpoint()'). Analyses
    can be adjusted for covariates using standard 'survival' package models.
    Ideal for biomarker analysis and patient stratification, where non-linear
    relationships may exist. Methods described in Yau, P.T.O. (2025)
    <doi:10.1101/2025.10.08.681246>.
License: GPL-3 + file LICENSE
     https://paytonyau.github.io/OptSurvCutR/
    cli,
    doParallel,
    foreach,
    doRNG,
    parallel,
    survival,
    survminer,
    srr
    broom,
    coxphf,
    patchwork,
    pROC,
    readxl,
    rgenoud,
    withr
    ropensci-review-tools/srr
Language: en-GB
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

	- [x] Regression and Supervised Learning *(Supervised method to define groups for survival regression models)*
	- [x] Exploratory Data Analysis (EDA) and Summary Statistics *(Identifying thresholds and visualising relationships)*

The target audience includes biostatisticians, bioinformaticians, clinical researchers, ecologists, and any scientist working with time-to-event data and continuous predictors. Scientific applications include biomarker discovery and validation, patient risk stratification in clinical studies, defining exposure thresholds in epidemiology or environmental science, and analysing dose-response relationships over time.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

`OptSurvCutR` represents both **an improvement on other R implementations** and the **first R implementation of a specific, statistically-driven 3-step workflow** for survival cut-point analysis.

Its primary innovations are:

1.  **Novel Cut-point Number Selection:** `OptSurvCutR` is the first R package to formalise selecting the *number* of cut-points as a model selection problem using **`AICc`** and **`BIC`**, in addition to `AIC`. This allows users to objectively compare models with 0, 1, 2, or more cut-points. This is a major advancement over tools like [`CutpointsOEHR`](https://cran.r-project.org/package=CutpointsOEHR), which is limited to `AIC` and focuses specifically on U or inverse-U shaped relationships with only a fixed, a 2-cut-point consideration.

2.  **Flexible Multi-Cut-point Optimisation:** The package is the first employs a genetic algorithm (via [`rgenoud`](https://cran.r-project.org/package=rgenoud)) to efficiently find the *location* of multiple (`>2`) cut-points simultaneously. This flexible approach can model complex, non-linear associations and is not limited to the single cut-point methods in packages like [`maxstat`](https://cran.r-project.org/package=maxstat) or [`survminer`](https://cran.r-project.org/package=survminer).

3.  **Complete, Validated Workflow:** `OptSurvCutR` is the first tool to integrate this novel number selection (Step 1) and flexible location finding (Step 2) with crucial downstream steps: covariate adjustment and bootstrap validation (Step 3). This complete, combined and validated workflow directly addresses the well-documented problems of inflated significance and overfitting in outcome-oriented cut-point selection (Altman et al., 1994; Faraggi & Simon, 1996). This rigorous approach aligns with guidance from health technology assessment (HTA), where understanding a model's assumptions and long-term plausibility is critical (Rutherford et al., 2020).

Yes. The package performs statistical analysis and includes example datasets (`crc_virome`, `germination`). The crc_virome data is derived from publicly available, anonymised TCGA data accessed via Smyth et al. (2024), and `germination` is simulated. The package itself does not collect, store, or transmit any user data or interact with human subjects directly.

-    What grade of badge are you aiming for? ([bronze, silver, gold](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges)) - **bronze**

-    If aiming for silver or gold, describe which of the [four aspects listed in the *Guide for Authors* chapter](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-silver) the package fulfils (at least one aspect for silver; three for gold) -
