+++ tests/testthat/test-core-functions.R
@@ -54,8 +54,8 @@ mock_data_skewed$predictor <- rexp(n_test, rate = 0.1)
 # Mock dataset with heavy censoring (90% censored)
 mock_data_heavy_censor <- mock_data
 mock_data_heavy_censor$event <- sample(0:1, n_test,
-                                       replace = TRUE,
-                                       prob = c(0.9, 0.1)
+  replace = TRUE,
+  prob = c(0.9, 0.1)
 )
 # Pathological dataset for bootstrap failure
 mock_data_pathological <- mock_data[1:20, ]
@@ -86,8 +86,8 @@ test_that("find_cutpoint_number genetic search works for all criteria", {
   # Test BIC (recommended default)
   res_bic <- suppressMessages(suppressWarnings(
     find_cutpoint_number(mock_data, "predictor", "time", "event",
-                         method = "genetic", criterion = "BIC", max_cuts = 2,
-                         maxiter = 5, nmin = 1
+      method = "genetic", criterion = "BIC", max_cuts = 2,
+      maxiter = 5, nmin = 1
     )
   ))
   skip_if(is.null(res_bic), "Genetic search returned NULL.")
@@ -98,8 +98,8 @@ test_that("find_cutpoint_number genetic search works for all criteria", {
   # Test AIC
   res_aic <- suppressMessages(suppressWarnings(
     find_cutpoint_number(mock_data, "predictor", "time", "event",
-                         method = "genetic", criterion = "AIC",
-                         maxiter = 5, nmin = 1
+      method = "genetic", criterion = "AIC",
+      maxiter = 5, nmin = 1
     )
   ))
   skip_if(is.null(res_aic), "Genetic search returned NULL.")
@@ -108,8 +108,8 @@ test_that("find_cutpoint_number genetic search works for all criteria", {
   # Test AICc
   res_aicc <- suppressMessages(suppressWarnings(
     find_cutpoint_number(mock_data, "predictor", "time", "event",
-                         method = "genetic", criterion = "AICc",
-                         maxiter = 5, nmin = 1
+      method = "genetic", criterion = "AICc",
+      maxiter = 5, nmin = 1
     )
   ))
   skip_if(is.null(res_aicc), "Genetic search returned NULL.")
@@ -122,8 +122,8 @@ test_that("find_cutpoint_number genetic search works for all criteria", {
 test_that("find_cutpoint_number systematic search works for BIC", {
   res_bic <- suppressMessages(suppressWarnings(
     find_cutpoint_number(mock_data, "predictor", "time", "event",
-                         method = "systematic", criterion = "BIC",
-                         max_cuts = 2, nmin = 1
+      method = "systematic", criterion = "BIC",
+      max_cuts = 2, nmin = 1
     )
   ))
   skip_if(is.null(res_bic), "Systematic search returned NULL.")
@@ -225,8 +225,8 @@ test_that("find_cutpoint_number handles too few unique values", {
 test_that("find_cutpoint_number handles AIC/AICc edge cases", {
   res_aic <- suppressMessages(suppressWarnings(
     find_cutpoint_number(tiny_data, "predictor", "time", "event",
-                         method = "systematic", criterion = "AIC",
-                         max_cuts = 1, nmin = 4
+      method = "systematic", criterion = "AIC",
+      max_cuts = 1, nmin = 4
     )
   ))
   skip_if(is.null(res_aic), "Systematic search returned NULL.")
@@ -236,8 +236,8 @@ test_that("find_cutpoint_number handles AIC/AICc edge cases", {
   )
[…]
