+++ R/OptSurvCutR-package.R
@@ -4,7 +4,3 @@
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
-
-## Register the @srrstats roxygen tag (no function is imported)
-#' @rawNamespace import(srr)
-NULL
