+++ docs/articles/index.md
@@ -1,8 +1,0 @@
-# Articles
-
-### All vignettes
-
-- [Stratifying Colorectal Cancer Patients by Gut Enterovirus
-  Abundance](https://paytonyau.github.io/OptSurvCutR/articles/CRC_Virome.md):
-- [Analysing Rapeseed Germination Data with
-  OptSurvCutR](https://paytonyau.github.io/OptSurvCutR/articles/Germination.md):

[…]
