### 2025-11-24 918b7cb
v0.1.9.2
see NEWS.md
