+++ R/find_cutpoint_number.R
@@ -8,53 +8,46 @@
 #'
 #' @section srrstats compliance:
 #' .
-#' @srrstats {RE4.1} Computes AIC, AICc, or BIC for model selection.
-#' @srrstats {G2.13} `cli_abort()` for invalid input.
-#' @srrstats {G2.14a} `NA` results handled via `na_result`.
-#' @srrstats {G3.1} `plot()` shows IC vs. number of cuts.
-#' @srrstats {G1.2} References provided for AIC, AICc, BIC.
+#' @srrstats {RE4.11} Computes Goodness-of-fit (AIC, AICc, BIC) for selection.
+#' @srrstats {G2.13} `cli_abort()` checks inputs pre-processing.
+#' @srrstats {G2.14} `NA` results handled via `na_result` helper.
+#' @srrstats {RE6.0} `plot()` method provided.
+#' @srrstats {RE6.2} `plot()` visualises model selection metric vs cuts.
+#' @srrstats {G1.0} References provided for AIC, AICc, BIC.
 #' @srrstats {G1.3} Systematic grid search (max_cuts <= 2) and
-#' `rgenoud` global optimization documented.
+#' `rgenoud` global optimisation documented.
 #' @srrstats {G1.5} Compared with `cutpointr`/`survminer` in vignette.
 #' @srrstats {G1.6} Numerical stability via `survival::coxph`
 #' and `rgenoud`; edge cases return `NA`.
 #' @srrstats {G2.3b} NSE via `as.formula()`/`data[...]`; no unsafe eval.
-#' @srrstats {G2.4} `NA` removed via `stats::na.omit()`;
-#' structured `NA` object returned on failure.
-#' @srrstats {G2.4e} `optimal_num_cuts`/`optimal_cuts` are `NA`
-#' when no valid solution found.
+#' @srrstats {G2.4} `NA` removed via `stats::na.omit()`.
+#' @srrstats {G5.2} `optimal_num_cuts` is `NA` when no valid solution found.
 #' @srrstats {G2.6} Input validation via direct checks.
 #' @srrstats {G2.8} Informative errors via `cli::cli_abort()`.
-#' @srrstats {G2.10} Warnings via `cli::cli_alert_warning()`.
-#' @srrstats {G2.12} Graceful degradation via `na_result()` for
+#' @srrstats {G5.2} Warnings via `cli::cli_alert_warning()`.
+#' @srrstats {G5.2} Graceful degradation via `na_result()` for
 #' empty data or model failures.
 #' @srrstats {G2.14c} `NA` propagation controlled.
-#' @srrstats {G4.0} All parameters/return values documented.
-#' @srrstats {G5.4c} Edge cases (zero rows, constant predictor) tested.
-#' @srrstats {G5.6a, G5.6b} Negative `max_cuts`/`nmin` rejected.
-#' @srrstats {G5.7} Large `max_cuts` constrained by `nmin`.
-#' @srrstats {G5.8d} `set.seed(seed)` for genetic reproducibility.
-#' @srrstats {G5.12} Systematic search scales poorly > 2 cuts.
+#' @srrstats {G1.4} All parameters/return values documented.
+#' @srrstats {RE4.17} `print()` method provided.
+#' @srrstats {RE4.18} `summary()` method provided.
 #'
 #' @srrstats {RE1.1} Assumes PH; check `summary()` for `cox.zph`.
-#' @srrstats {RE1.2} Provides AIC, AICc, BIC fit statistics.
-#' @srrstats {RE1.3, RE1.3a} PH diagnostics via `cox.zph` in `summary()`.
[…]
