+++ R/srr-stats-standards.R
@@ -1,0 +1,8 @@
+#' NA_standards
+#'
+#' Any non-applicable standards can have their tags changed from `@srrstatsTODO`
+#' to `@srrstatsNA`, and placed together in this block, along with explanations
+#' for why each of these standards have been deemed not applicable.
+#' (These comments may also be deleted at any time.)
+#' @noRd
+NULL
