### 2025-11-25 #731 @paytonyau
Thanks for the review @mpadge and for allowing the submission to proceed despite the missing pre-submission step—I appreciate that.

I have addressed the points raised above:

I have updated the documentation to use @srrstatsNA for all non-applicable standards.
I have reviewed the full list of standards to ensure the descriptions align with the correct standard numbers (including the mismatches flagged in G2.15 and G2.16).
I am now triggering the checks as requested:
