### 2026-03-15 e9987c3
Add graphical abstract and docs updates
Add graphical abstract image and polish documentation/site configuration. This commit: adds image/graphical_abstract.PNG and embeds it in README; updates NEWS with release notes and presentation entries; modifies README examples (adds dplyr, replaces plot_schoenfeld with plot_cutpoint_residuals, reformats example plots, removes BuyMeACoffee section, and adds protocol link); updates _pkgdown.yml (syntax highlighting, repo config, sidebar/navbar links and icons); removes vignettes/.gitignore; and updates vignettes (use plot_cutpoint_residuals in CRC_Virome
[…]
