+++ README.md
@@ -9,15 +9,19 @@
 
 `OptSurvCutR` (**Opt**imal **Surv**ival **Cut**-points **R**) provides a rigorous, reproducible **three-step workflow** for discovering the optimal number and location of cut-points in time-to-event (survival) data. Designed for continuous predictors (e.g., gene expression, virome abundance, biomarkers), it moves beyond arbitrary median splits to fully **data-driven stratification**.
 
+<p align="center">
+  <img src="image/graphical_abstract.png" alt="OptSurvCutR Graphical Abstract" width="100%">
+</p>
+
 ## Why OptSurvCutR?
 
-| Feature                        | Benefit                                                               |
+| Feature                       | Benefit                                                               |
 |-------------------------------|-----------------------------------------------------------------------|
 | Optimal number of cuts        | Uses AIC, AICc, or BIC to select 0–k cut-points                       |
-| Flexible search               | Systematic grid or genetic algorithm (`rgenoud`)                     |
+| Flexible search               | Systematic grid or genetic algorithm (`rgenoud`)                      |
 | Covariate adjustment          | Control for confounders during cut-point discovery                    |
-| Bootstrap validation          | 95% confidence intervals for cut-point stability                     |
-| Publication-ready plots       | Kaplan–Meier, optimisation curves, Schoenfeld residual diagnostics   |
+| Bootstrap validation          | 95% confidence intervals for cut-point stability                      |
+| Publication-ready plots       | Kaplan–Meier, optimisation curves, Schoenfeld residual diagnostics    |
 
 ## Installation
 You can install the development version of `OptSurvCutR` from GitHub. Note that the genetic algorithm (`method = "genetic"`) requires the `rgenoud` package, which should be installed separately from CRAN if you plan to use it.
@@ -37,6 +41,7 @@ remotes::install_github("paytonyau/OptSurvCutR")
 ```r
 library(OptSurvCutR)
 library(survival)
+library(dplyr)
 
 data("crc_virome")
 
@@ -64,10 +69,10 @@ val <- validate_cutpoint(cuts, num_replicates = 200, seed = 456)
 summary(val)
 
 # Step 4: Visualise results
-plot(cuts, type = "outcome")        # Kaplan–Meier curves
-plot(cuts, type = "distribution")  # Predictor + cuts
-plot_schoenfeld(cuts)               # Proportional hazards check
-plot(val)                           # Bootstrap density + 95% CI
+plot(cuts, type = "outcome")          # Kaplan–Meier curves
+plot(cuts, type = "distribution")     # Predictor + cuts
+plot_cutpoint_residuals(cuts)         # Proportional hazards check (Schoenfeld)
+plot(val)                             # Bootstrap density + 95% CI
 ```
 
 ## Workflow Summary
@@ -76,8 +81,10 @@ OptSurvCutR provides a three-step workflow for cut-point analysis:
 1.  `find_cutpoint_number()`: – selects the statistically optimal number of cut-points using information criteria
 2.  `find_cutpoint()`: locates exact cut-point values (systematic or genetic search)
 3.  `validate_cutpoint()`: assesses stability via bootstrapping with 95% confidence intervals
+
 ## Resources
 - **Vignettes**: `browseVignettes("OptSurvCutR")`
+- **Protocol**: https://paytonyau.github.io/OptSurvCutR/index.html
 - **Package Website**:  https://paytonyau.github.io/OptSurvCutR/
[…]
