### 2026-03-15 60bcc4a
Revert "Add graphical abstract and docs updates"
This reverts commit e9987c3f275c8d07ee5de460e9aae696bf5a4b38.
