### 2026-03-15 1d1a0ad
v.0.1.9.3
Add a new graphical abstract (image/graphical_abstract.jpg) and include it in README.md; update NEWS.md with a 0.1.9.3 entry describing the addition and recent presentations. Update .github/workflows/R-CMD-check.yaml to remove the oldrel-1 runner to avoid a ggrepel >= 4.5.0 dependency conflict and set FORCE_JAVASCRIPT_ACTIONS_TO_NODE24 to silence a Node.js 20 warning.
