### 2026-03-22 5b6997f
v0.1.9.3
Bump package version to 0.1.9.3 (DESCRIPTION) and prepare release/site rebuild. Update NEWS and docs/news with the graphical abstract and presentation entries (FOSDEM, R!SK), embed the graphical abstract in index pages (docs/index.html, docs/index.md, docs/llms.txt), and update pkgdown metadata (last_built). Regenerate many pkgdown-generated HTML files to show the new released version and updated vignette sessionInfo (R and package versions). Remove .Rhistory and update search.json accordingly.
