+++ docs/llms.txt
@@ -7,6 +7,10 @@ data. Designed for continuous predictors (e.g., gene expression, virome
 abundance, biomarkers), it moves beyond arbitrary median splits to fully
 **data-driven stratification**.
 
+![OptSurvCutR Graphical Abstract](image/graphical_abstract.jpg)
+
+OptSurvCutR Graphical Abstract
+
 ## Why OptSurvCutR?
 
 | Feature | Benefit |
