### 2026-04-07 90ebbda
v0.1.9.3 fix _pkgdown.yml
Update navbar and sidebar vignette hrefs to point to the new article filenames (articles/CRC_Virome.html and articles/Germination.html). Remove an obsolete FIX comment and normalize indentation in the components/vignettes section for readability.
