### 2026-04-09 5c38803
v1.9.3
Various packaging and documentation cleanups: update GitHub Actions matrix (add oldrel-1, remove Node.js warning flag) and normalize build_args formatting; simplify DESCRIPTION license field to "GPL-3" and remove the separate LICENSE file; update README badge URL and point graphical abstract to man/figures; move image, adjust R and Rd documentation to use consistent DOI formatting; remove tests/spelling.R; and add several entries to inst/WORDLIST. These changes tidy metadata, fix doc paths/formatting, and streamline CI/tests.
