+++ man/germination.Rd
@@ -16,8 +16,7 @@ A data frame with 1040 rows and 5 variables:
 }
 \source{
 Computationally simulated based on parameters derived from
-  Haj Sghaier et al. (2022) [\doi{10.3390/plants11212819}].
-  \url{https://www.mdpi.com/2223-7747/11/21/2819}
+  Haj Sghaier et al. (2022) \doi{10.3390/plants11212819}.
 }
 \description{
 Simulated germination and early growth of rapeseed (Brassica napus L.)
@@ -28,7 +27,7 @@ Haj Sghaier et al. (2022).
 Haj Sghaier, A., Tarnawa, Á., Khaeim, H., Kovács, G. P.,
 Gyuricza, C., & Kende, Z. (2022). The Effects of Temperature and
 Water on the Seed Germination and Seedling Development of Rapeseed
-(Brassica napus L.). *Plants*,11(21), 2819.
-https://doi.org/10.3390/plants11212819
+(Brassica napus L.). *Plants*, 11(21), 2819.
+\doi{10.3390/plants11212819}
 }
 \keyword{datasets}
