+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: Provides a robust 3-step workflow for optimal cut-point analysis in
     and validate stability via bootstrapping ('validate_cutpoint()'). Analyses
     can be adjusted for covariates using standard 'survival' package models.
     Ideal for biomarker analysis and patient stratification.
-License: GPL-3 + file LICENSE
+License: GPL-3
 URL: https://github.com/paytonyau/OptSurvCutR,
      https://paytonyau.github.io/OptSurvCutR/
 BugReports: https://github.com/paytonyau/OptSurvCutR/issues
