### 2026-04-09 8fca69c
v0.2.0 CRAN submission
Bump package version to 0.2.0 and add Date. Add R-hub GitHub Actions workflow (rhub.yaml) for platform checks. Mark this as the initial CRAN submission in NEWS.md and cran-comments.md. Update pkgdown configuration and site files (nav links, version text, search/sitemap, reference pages) to reflect new vignette filenames and version. Remove generated license pages from docs and add a graphical abstract image to docs/reference/figures. Miscellaneous documentation and reference updates to prepare the package for CRAN.
