+++ cran-comments.md
@@ -1,5 +1,13 @@
-## R CMD check results
+## OptSurvCutR 0.2.0
 
+This is a new submission. 
+
+## Test environments
+* local Windows 11 install, R 4.5.3
+* win-builder (devel and release)
+
+## R CMD check results
 0 errors | 0 warnings | 1 note
 
-* This is a new release.
+* Note: "New submission" - This is expected as this is the first submission of this package.
+* Note: "unable to verify current time" - Harmless network timeout.
