### 2026-04-20 7c56795
Bump to 0.2.1: CRAN fixes and branding
Release bump to 0.2.1 (2026-04-20) with multiple CRAN compliance and polish changes.

- DESCRIPTION: version/date updated, removed single quotes around package/function names and added Config/Needs/website entry for logo.
- R code: main S3 functions (find_cutpoint, find_cutpoint_number, validate_cutpoint) now return results instead of forcing print()/invisible() (silences output when assigned).
- Branding: bundled package logo (man/figures/logo.png) and added favicons/site manifest for pkgdown and docs; _pkgdown.yml updated to reference logo.
- Docs & REA
[…]
