+++ README.md
@@ -7,7 +7,11 @@
 [![License: GPL-3](https://img.shields.io/badge/License-GPL%203-blue.svg)](https://www.gnu.org/licenses/gpl-3.0.html)
 <!-- badges: end -->
 
+<img src="man/figures/logo.png" align="right" height="139" alt="OptSurvCutR logo" />
+
+
 `OptSurvCutR` (**Opt**imal **Surv**ival **Cut**-points **R**) provides a rigorous, reproducible **three-step workflow** for discovering the optimal number and location of cut-points in time-to-event (survival) data. Designed for continuous predictors (e.g., gene expression, virome abundance, biomarkers), it moves beyond arbitrary median splits to fully **data-driven stratification**.
+
 
 ![OptSurvCutR Graphical Abstract](man/figures/graphical_abstract.jpg)
 
