+++ man/OptSurvCutR-package.Rd
@@ -6,7 +6,9 @@
 \alias{OptSurvCutR-package}
 \title{OptSurvCutR: Optimal Survival Cut-Point Discovery for Time-to-Event Analysis}
 \description{
-Provides a robust 3-step workflow for optimal cut-point analysis in time-to-event (survival) data. Functions determine the optimal number of cut-points ('find_cutpoint_number()'), find their precise locations ('find_cutpoint()') using systematic or genetic algorithms (via 'rgenoud'), and validate stability via bootstrapping ('validate_cutpoint()'). Analyses can be adjusted for covariates using standard 'survival' package models. Ideal for biomarker analysis and patient stratification.
+\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
+
+Provides a robust 3-step workflow for optimal cut-point analysis in time-to-event (survival) data. Functions determine the optimal number of cut-points (find_cutpoint_number()), find their precise locations (find_cutpoint()) using systematic or genetic algorithms (via rgenoud), and validate stability via bootstrapping (validate_cutpoint()). Analyses can be adjusted for covariates using standard survival package models. Ideal for biomarker analysis and patient stratification.
 }
 \seealso{
 Useful links:
