### 2026-05-03 #731 @maurolepore
Following up on @emilyriederer's note from March, the package looks good and is ready for review; the next step is finding a handling editor. I'll work on that and follow up here.

Thanks for your patience and for sharing your work with the rOpenSci community.
