### 2026-05-03 #731 @maurolepore
@paytonyau

I looked at the thread more closely: @emilyriederer and @mpadge agree this package is in scope and in good shape. They guided you to update the template to mark this issue as the pre-submission inquiry.

Since the pre-submission has served its purpose, I'll go ahead and close it.

Next, please open a new issue using the [statistical software submission template](https://github.com/ropensci/software-review/issues/new?template=F-submit-statistical-software-for-review.md), referencing this one (#731) in the pre-submission inquiry checkbox.

Emily already flagged some delay in staffing stats editors. I'll keep working on that. Thanks for your patience.
